Release 0.01
22.10.2024
The "Find GSI" button has been renamed to "Find GSI system", as well as "Phone Status" to "Phone Visibility". Renamed the heading "Check status via" to "Check Visibility via"
The new button is "Forum in 4PDA". It opens a link with the 4PDA forum about the official firmware. Along with it, a new parameter is added in the config, which gives a link to the Forum
The new button is "Forum in 4PDA". It opens a link with the 4PDA forum about unofficial firmware. Along with it, a new parameter is added in the config, which gives a link to the Forum

23.10.2024 
Added the "Flash custom recovery/boot.img" button

25.10.2024
A frame has been created for twrp and boot firmware

26.10.2024
The frame and the frame functions are fully functional
Standardization of configs and parameters

27.10.2024
The "donations" frame has been added

31.10.2021
Fixed a bug that bricks half of the program on windows. Accordingly, the images will be loaded